+++ stubs/Doctrine/Bundle/MongoDBBundle/Repository/ServiceDocumentRepositoryInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Doctrine\Bundle\MongoDBBundle\Repository;
+
+interface ServiceDocumentRepositoryInterface
+{
+}
