+++ tests/Rules/Doctrine/NoGetRepositoryOnServiceRepositoryEntityRule/Source/Entity/NormalEntity.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity(repositoryClass=\Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source\Repository\NormalRepository::class)
+ */
+class NormalEntity
+{
+}
