--- tests/Rules/Doctrine/NoGetRepositoryOnServiceRepositoryEntityRule/Source/NormalRepository.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source;
-
-use Doctrine\ORM\EntityRepository;
-
-final class NormalRepository extends EntityRepository
-{
-}
