+++ tests/Rules/Doctrine/NoGetRepositoryOnServiceRepositoryEntityRule/Source/Repository/NormalRepository.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+final class NormalRepository extends EntityRepository
+{
+}
