+++ tests/Rules/Doctrine/NoGetRepositoryOnServiceRepositoryEntityRule/Source/Repository/RepositoryWithInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source\Repository;
+
+use Doctrine\Bundle\MongoDBBundle\Repository\ServiceDocumentRepositoryInterface;
+
+final class RepositoryWithInterface implements ServiceDocumentRepositoryInterface
+{
+}
