+++ tests/Rules/Doctrine/NoGetRepositoryOnServiceRepositoryEntityRule/Source/Repository/SomeServiceRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source;
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOnServiceRepositoryEntityRule\Source\Repository;
 
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
